Move bootstrap witness tests from `BalanceSpec` to `SignSpec`
